refactor: resolve duplicate backend directories by merging src/backend into backend
- Ported /health endpoint, 100kb JSON limit, and 404 handler to backend/app.js
- Ported peer matching routes and controllers (matchRoutes.js, matchController.js)
- Merged gamification and study partner fields into backend/models/User.js
- Added email/URL validation to backend/utils/sendEmail.js
- Deleted orphaned src/backend directory to reduce confusion and Vite bundle size
