+++ backend/utils/sendEmail.js
@@ -1,7 +1,32 @@
 import nodemailer from "nodemailer";
 import { env } from "../config.js";
 
+// Basic email format check — catches the most common malformed inputs
+// before a network call is made.
+const isValidEmail = (value) => /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(String(value));
+
+// Ensure the reset URL uses a safe scheme before embedding it in an email.
+// A javascript: or data: URL would execute in some email clients.
+const isSafeUrl = (value) => {
+  try {
+    const parsed = new URL(String(value));
+    return parsed.protocol === "https:" || parsed.protocol === "http:";
+  } catch {
+    return false;
+  }
+};
+
 export const sendEmail = async (email, url) => {
+  if (!isValidEmail(email)) {
+    throw new Error("sendEmail: the recipient email address is not valid.");
+  }
+
+  if (!isSafeUrl(url)) {
+    throw new Error(
+      "sendEmail: the reset URL must be a valid http or https URL."
+    );
+  }
+
   const emailUser = env.EMAIL_USER;
   const emailPass = env.EMAIL_PASS;
