--- src/backend/controllers/authController.js
@@ -1,133 +1,0 @@
-import bcrypt from "bcryptjs";
-import crypto from "crypto";
-import User from "../models/User.js";
-import { sendEmail } from "../utils/sendEmail.js";
-
-const RESET_TOKEN_TTL_MS = 15 * 60 * 1000;
-const GENERIC_RESET_MESSAGE =
-  "If an account with that email exists, a password reset link has been sent.";
-
-const buildFrontendBaseUrl = (req) => {
-  const configuredBaseUrl =
-    process.env.PASSWORD_RESET_BASE_URL ||
-    process.env.FRONTEND_URL ||
-    process.env.CLIENT_URL;
-
-  if (configuredBaseUrl) {
-    return configuredBaseUrl;
-  }
-
-  const protocol = req.headers["x-forwarded-proto"] || req.protocol;
-  return `${protocol}://${req.get("host")}`;
-};
-
-export const forgotPassword = async (req, res) => {
-  try {
-    const { email } = req.body || {};
-
-    if (!email || typeof email !== "string") {
-      return res.status(400).json({
-        success: false,
-        message: "Email is required",
-      });
-    }
-
-    const normalizedEmail = String(email).trim().toLowerCase();
-    const user = await User.findOne({ email: normalizedEmail });
-
-    if (!user) {
-      return res.status(200).json({
-        success: true,
-        message: GENERIC_RESET_MESSAGE,
-      });
-    }
-
-    const rawResetToken = crypto.randomBytes(32).toString("hex");
-    const hashedResetToken = crypto
-      .createHash("sha256")
-      .update(rawResetToken)
-      .digest("hex");
-
-    user.resetPasswordToken = hashedResetToken;
-    user.resetPasswordExpire = new Date(Date.now() + RESET_TOKEN_TTL_MS);
-    await user.save({ validateBeforeSave: false });
-
-    const frontendBaseUrl = buildFrontendBaseUrl(req);
-    const resetUrl = `${frontendBaseUrl.replace(/\/$/, "")}/reset-password?token=${rawResetToken}`;
-
-    try {
-      await sendEmail(user.email, resetUrl);
-    } catch (mailError) {
-      user.resetPasswordToken = undefined;
-      user.resetPasswordExpire = undefined;
-      await user.save({ validateBeforeSave: false });
-
-      return res.status(500).json({
-        success: false,
-        message: "Unable to send reset email. Please try again later.",
-      });
-    }
-
-    res.status(200).json({
-      success: true,
-      message: GENERIC_RESET_MESSAGE,
-    });
-  } catch (error) {
-    console.error("Forgot password error:", error);
-    res.status(500).json({
-      success: false,
-      message: error.message || "Server Error",
-    });
-  }
-};
-
-export const resetPassword = async (req, res) => {
-  try {
-    const { token } = req.params;
-    const password = req.body?.password || req.body?.newPassword;
-
-    if (!token || !password) {
-      return res.status(400).json({
-        success: false,
-        message: "Token and new password are required",
-      });
-    }
-
-    if (String(password).length < 6) {
-      return res.status(400).json({
-        success: false,
-        message: "Password must be at least 6 characters long",
-      });
-    }
-
-    const hashedToken = crypto.createHash("sha256").update(token).digest("hex");
-
-    const user = await User.findOne({
-      resetPasswordToken: hashedToken,
-      resetPasswordExpire: { $gt: Date.now() },
-    });
-
-    if (!user) {
-      return res.status(400).json({
-        success: false,
-        message: "Reset token is invalid or has expired",
-      });
-    }
-
-    user.password = await bcrypt.hash(password, 10);
-    user.resetPasswordToken = undefined;
-    user.resetPasswordExpire = undefined;
-    await user.save();
-
-    res.status(200).json({
-      success: true,
-      message: "Password has been reset successfully",
-    });
-  } catch (error) {
-    console.error("Reset password error:", error);
-    res.status(500).json({
-      success: false,
-      message: error.message || "Server Error",
-    });
-  }
-};
