--- src/backend/middlewares/requireAuth.js
@@ -1,79 +1,0 @@
-import { createClient } from "@supabase/supabase-js";
-
-const supabaseAdmin = createClient(
-  process.env.SUPABASE_URL,
-  process.env.SUPABASE_SERVICE_ROLE_KEY || process.env.SUPABASE_ANON_KEY
-);
-
-/**
- * Express middleware that validates a Supabase JWT from the Authorization header.
- * Rejects requests with no token or an invalid/expired token with 401.
- * Attaches the authenticated user object to req.user on success.
- */
-export const requireAuth = async (req, res, next) => {
-  const authHeader = req.headers.authorization;
-
-  if (!authHeader || !authHeader.startsWith("Bearer ")) {
-    return res.status(401).json({ error: "Authentication required" });
-  }
-
-  const token = authHeader.slice(7);
-
-  const { data, error } = await supabaseAdmin.auth.getUser(token);
-
-  if (error || !data?.user) {
-    return res.status(401).json({ error: "Invalid or expired session" });
-  }
-
-  req.user = data.user;
-  next();
-};
-
-const deriveActiveRoles = (profile) => {
-  const roles = [];
-
-  if (profile?.is_mentor) {
-    roles.push("mentor");
-  }
-
-  if (profile?.is_learner) {
-    roles.push("learner");
-  }
-
-  return roles;
-};
-
-export const requireProfileRole = (...allowedRoles) => async (req, res, next) => {
-  try {
-    if (!req.user?.id) {
-      return res.status(401).json({ error: "Authentication required" });
-    }
-
-    const { data: profile, error } = await supabaseAdmin
-      .from("profiles")
-      .select("id, is_mentor, is_learner")
-      .eq("id", req.user.id)
-      .maybeSingle();
-
-    if (error) {
-      console.error("Profile authorization error:", error);
-      return res.status(500).json({ error: "Unable to verify account permissions" });
-    }
-
-    if (!profile) {
-      return res.status(403).json({ error: "Not authorized to access this resource" });
-    }
-
-    const activeRoles = deriveActiveRoles(profile);
-    if (allowedRoles.length > 0 && !allowedRoles.some((role) => activeRoles.includes(role))) {
-      return res.status(403).json({ error: "Not authorized to access this resource" });
-    }
-
-    req.profile = profile;
-    req.roles = activeRoles;
-    next();
-  } catch (error) {
-    console.error("Profile authorization error:", error);
-    res.status(500).json({ error: "Unable to verify account permissions" });
-  }
-};
